% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_table.R
\name{base_table}
\alias{base_table}
\title{Return base table of values}
\usage{
base_table(lb, x1, wb, b, kb, qthresh, qmean, por)
}
\arguments{
\item{lb}{BASIN LENGTH}

\item{x1}{LONGITUDINAL SCALING PARAMETER}

\item{wb}{WIDTH OF BASE ZONE}

\item{b}{EXPONENT FOR SATURATED HEIGHT AS A FUNCTION OF LENGTH:Z=\link{(xb-X0)/x1}^BETA}

\item{kb}{HORIZONAL HYDRAULIC CONDUCTIVITY OF BASE}

\item{qthresh}{qthresh}

\item{qmean}{qmean}

\item{por}{DRAINABLE POROSITY}
}
\value{
a dataframe
}
\description{
\code{base_table()} returns a lookup table for base with saturated height,
intersection, hydraulic gradient, storage and discharge
Creates a table with Zb, hydraulic gradient (dZb/dx), Sb, and Qb for 1202 values of Xb from 0 to Lb
}
\examples{
base_table(1,2,3,4,5,6,7,8)
}
