% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin_char.R
\name{basin_char}
\alias{basin_char}
\title{Get basin characteristics for a USGS gage}
\usage{
basin_char(siteNumber)
}
\arguments{
\item{siteNumber}{character USGS gage ID}
}
\value{
vector with basin characteristics
}
\description{
Get basin characteristics for a USGS gage
}
\examples{
basin_char("12189500")
}
