% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_ci.R
\name{bf_ci}
\alias{bf_ci}
\title{Calculate prediction intervals}
\usage{
bf_ci(bf_mod_out, return_type = "ci")
}
\arguments{
\item{bf_mod_out}{bfmodout}

\item{return_type}{type of return}
}
\value{
a dataframe
}
\description{
Calculate credible prediction intervals for streamflow
}
