% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bf_params_usgs}
\alias{bf_params_usgs}
\title{USGS gage parameters data}
\format{
\subsection{\code{bf_params_usgs}}{

A data frame with 13,208 rows and 18 columns:
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\source{
\url{https://www.sciencebase.gov/catalog/item/5f90ef0282ce720ee2d29b7e}
}
\usage{
bf_params_usgs
}
\description{
A set of calibration numbers for USGS gages
Report ...
}
\keyword{datasets}
