% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_sep.R
\name{bf_sep}
\alias{bf_sep}
\title{Run baseflow separation model}
\usage{
bf_sep(qin, dy, timestep, error_basis, basin_char, gw_hyd, flow)
}
\arguments{
\item{qin}{a numeric vector of streamflow volume for each time step}

\item{dy}{date or POSIXlt vector of date-times for each time step}

\item{timestep}{'day' or 'hour'}

\item{error_basis}{'base' or 'total'}

\item{basin_char}{vector with 6 parameters for basin characteristics}

\item{gw_hyd}{vector with 6 parameters for groundwater hydraulics}

\item{flow}{vector with 6 parameters for streamflow}
}
\value{
a dataframe
}
\description{
HYDROGRAPH IS SEPARATED INTO THREE COMPONENTS:
\enumerate{
\item DIRECT RUNOFF OF PPT IMPULSE ON SATURATED SURFACE (SATURATED OVERLAND FLOW) AND PPT-INFILTRATION (INFILTRATION LIMITED RUNOFF)
\item SUBSURFACE FLOW LATERALLY THROUGH SURFACE ZONE, UNIFORM SATURATED THICKNESS AND HYDRAULIC GRADIENT
\item BASE FLOW LONGITUDINALLY THROUGH SUBSURFACE ZONE, VARIABLE SATURATED THICKNESS AND HYDRAULIC GRADIENT DEPENDING ON WATER LEVEL
}
}
