% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_q.R
\name{dir_q}
\alias{dir_q}
\title{Direct runoff}
\usage{
dir_q(lb, a, z, i)
}
\arguments{
\item{lb}{BASIN LENGTH}

\item{a}{HYDRAULIC GRADIENT FOR SURFACE (CONSTANT)}

\item{z}{WATER SURFACE ELEVATION AT SURFACE OF SURFACE ZONE (DATUM IS CHANNEL)}

\item{i}{IMPULSE DEPTH}
}
\value{
a numeric scalar
}
\description{
\code{dir_q()} calculates the direct runoff
}
\examples{
dir_q(1,2,3,4)
}
