% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.R
\name{flow}
\alias{flow}
\title{Get flow parameters for a USGS gage}
\usage{
flow(siteNumber)
}
\arguments{
\item{siteNumber}{USGS gage ID}
}
\value{
a vector of flow parameters
}
\description{
Get flow parameters for a USGS gage
}
\examples{
flow("12189500")
}
