% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_hyd.R
\name{gw_hyd}
\alias{gw_hyd}
\title{Get groundwater hydrology for a USGS gage}
\usage{
gw_hyd(siteNumber)
}
\arguments{
\item{siteNumber}{USGS gage ID}
}
\value{
a vector of groundwater hydrology
}
\description{
Get groundwater hydrology for a USGS gage
}
\examples{
gw_hyd("12189500")
}
