% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recharge.R
\name{recharge}
\alias{recharge}
\title{Calculate recharge}
\usage{
recharge(lb, xb, ws, kz, zs, por)
}
\arguments{
\item{lb}{BASIN LENGTH}

\item{xb}{LONGITUDINAL LOCATION OF BASE WATER LEVEL INTERSECTION WITH SURFACE}

\item{ws}{WIDTH OF SURACE ZONE (ONE SIDE OF CHANNEL)}

\item{kz}{VERTICAL HYDRAULIC CONDUCTIVITY}

\item{zs}{SATURATED THICKNESS OF SURFACE}

\item{por}{DRAINABLE POROSITY}
}
\value{
a numeric scalar
}
\description{
\code{recharge()} calculates the portion of water stored in the surface reservoir that recharges the base reservoir, R
}
\examples{
recharge(1,2,3,4,5,6)
}
