% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sur_q.R
\name{sur_q}
\alias{sur_q}
\title{Discharge for surface zone}
\usage{
sur_q(lb, a, ks, z)
}
\arguments{
\item{lb}{BASIN LENGTH}

\item{a}{HYDRAULIC GRADIENT FOR SURFACE (CONSTANT)}

\item{ks}{HYDRAULIC CONDUCTIVITY OF SURFACE}

\item{z}{WATER SURFACE ELEVATION AT SURFACE OF SURFACE ZONE (DATUM IS CHANNEL)}
}
\value{
a numeric scalar
}
\description{
\code{sur_q()} calculates discharge from the surface reservoir, Qs, given Zs
}
\examples{
sur_q(1,2,3,4)
}
