% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sur_store.R
\name{sur_store}
\alias{sur_store}
\title{Calculate Surface Storage}
\usage{
sur_store(lb, a, ws, por, zs)
}
\arguments{
\item{lb}{BASIN LENGTH}

\item{a}{HYDRAULIC GRADIENT FOR SURFACE (CONSTANT)}

\item{ws}{WIDTH OF SURACE ZONE (ONE SIDE OF CHANNEL)}

\item{por}{DRAINABLE POROSITY}

\item{zs}{SATURATED THICKNESS OF SURFACE}
}
\value{
a numeric scalar
}
\description{
\code{sur_store()} calculates surface storage as the difference between the
total surface zone volume and the unsaturated portion below Z (surface saturation thickness),
both are triangular prisms
}
\examples{
sur_store(1,2,3,4,5)
}
