% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sur_z.R
\name{sur_z}
\alias{sur_z}
\title{Calculate saturated thickness in surface zone}
\usage{
sur_z(lb, a, ws, por, ss)
}
\arguments{
\item{lb}{BASIN LENGTH}

\item{a}{HYDRAULIC GRADIENT FOR SURFACE (CONSTANT)}

\item{ws}{WIDTH OF SURACE ZONE (ONE SIDE OF CHANNEL)}

\item{por}{DRAINABLE POROSITY}

\item{ss}{unknown}
}
\value{
a numeric scalar
}
\description{
Calculates the saturated thickness of the surface reservoir, Zs, given the volume of water stored in the surface reservoir, Ss.
}
\examples{
sur_z(1,2,3,4,5)
}
